embed the env config for integration tests within the setup and not at the project level
