+++ tests/integration/env.offline
@@ -1,0 +1,8 @@
+# Used for integration tests
+AE5_HOSTNAME=0.0.0.0
+AE5_USERNAME=mock-user
+AE5_PASSWORD=mock-password
+AE5_ADMIN_USERNAME=mock-adm-user
+AE5_ADMIN_PASSWORD=mock-adm-user-password
+AE5_K8S_ENDPOINT=ssh:centos
+AE5_K8S_PORT=23456
