+++ tests/integration/test_api.py
@@ -1,35 +1,10 @@
 import json
-import os
 import subprocess
 import shlex
 from typing import Dict, List
 import uuid
 
-import requests
-
-
-def reset_mock_state() -> bool:
-    response = requests.delete(url=f"https://{os.environ['AE5_HOSTNAME']}/mock/state", verify=False)
-    if response.status_code != 200:
-        message: str = f"reset_mock_state saw: {response.status_code}, text: {response.text}"
-        raise Exception(message)
-    return True
-
-
-def set_mock_state(mock_state: Dict) -> Dict:
-    response = requests.patch(url=f"https://{os.environ['AE5_HOSTNAME']}/mock/state", json=mock_state, verify=False)
-    if response.status_code != 200:
-        message: str = f"set_mock_state saw: {response.status_code}, text: {response.text}"
-        raise Exception(message)
-    return response.json()
-
-
-def get_mock_state() -> Dict:
-    response = requests.get(url=f"https://{os.environ['AE5_HOSTNAME']}/mock/state", verify=False)
-    if response.status_code != 200:
-        message: str = f"get_mock_state saw: {response.status_code}, text: {response.text}"
-        raise Exception(message)
-    return response.json()
+from tests.integration.mock.ae5 import AE5MockClient
 
 
 def shell_out(cmd: str) -> tuple[str, str, int]:
@@ -164,8 +139,8 @@ def test_user_list():
     ]
 
     for test_case in test_cases:
-        reset_mock_state()
-        scenario = set_mock_state(mock_state=test_case["state"])
+        AE5MockClient.reset_mock_state()
+        scenario = AE5MockClient.set_mock_state(mock_state=test_case["state"])
         print(scenario)
         outs, errs, returncode = shell_out(cmd=test_case["command"])
